requirements.txt
fastapi~=0.116.1
uvicorn
langchain~=0.3.27
langchain-community~=0.3.27
langchain-google-genai~=2.1.8
chromadb
PyPDF2~=3.0.1
unstructured
bs4
requests
duckduckgo-search~=8.1.1
newspaper3k~=0.2.8
wikipedia~=1.4.0
python-multipart
langchain-core~=0.3.72
pydantic~=2.11.7
protobuf~=6.31.1
python-dotenv~=1.1.1
starlette~=0.47.2
ddgs~=9.5.0
urllib3~=2.5.0
sentence-transformers~=5.0.0